Small cleanup.

+++ src/bitstream/DefaultInputBitStream.hpp
@@ -105,7 +105,7 @@ namespace kanzi {
 
            if (_position + 7 > _maxPosition) {
                // End of stream: overshoot max position => adjust bit index
-               uint shift = (_maxPosition - _position) << 3;
+               uint shift = uint(_maxPosition - _position) << 3;
                _availBits = shift + 8;
                uint64 val = 0;
